feat(projects): Allow assigning users to projects
This commit introduces the ability to assign a user to a project during
creation and editing.

The ProjectController now provides a list of all users to the create and
edit views. These views now feature a dropdown menu, allowing the person
creating or editing the project to select an assignee.